+++ server/database/migrations/2021_04_14_152647_create_assignacio_balls_table.php
@@ -14,9 +14,9 @@ class CreateAssignacioBallsTable extends Migration
     public function up()
     {
         Schema::create('assignacio_balls', function (Blueprint $table) {
-            $table->integer('persona_id');
+            $table->unsignedBigInteger('persona_id');
             $table->foreign('persona_id')->references('id')->on('personas');
-            $table->integer('ball_id');
+            $table->unsignedBigInteger('ball_id');
             $table->foreign('ball_id')->references('id')->on('tipus_balls');
             $table->timestamps();
         });
