+++ server/database/migrations/2021_04_14_152711_create_diplomas_table.php
@@ -16,7 +16,7 @@ class CreateDiplomasTable extends Migration
         Schema::create('diplomas', function (Blueprint $table) {
             $table->id();
             $table->timestamps();
-            $table->integer('usuari_id');
+            $table->unsignedBigInteger('usuari_id');
             $table->foreign('usuari_id')->references('id')->on('usuaris');
             $table->string('titol', 100);
         });
